csweb/app/requirements.txt
flask>="1.1.2"
flask-sqlalchemy>="2.4.4"
sqlalchemy>="1.3.22"
pymysql>= "1.0.2"
flask-migrate>="2.6.0"
flask-login>="0.5.0"
werkzeug>="1.0.1"
sendgrid>="6.5.0"
cryptography>="3.3.1"
boto3>="1.12.39"
python-dotenv>="0.15.0"
boto3>="1.12.39"
python-dotenv>="0.15.0"